requirements.txt
autogen-agentchat
autogen-ext[magentic-one,openai]
azure-identity
python-dotenv
colorama
azure.search.documents
azure.ai.inference
aiohttp
chainlit